+++ tests/test_notifications.py
@@ -1,9 +1,12 @@
 from os import path
 from unittest.mock import MagicMock, patch
 
-import airflow
+from dbt_airflow_factory.constants import (
+    IS_AIRFLOW_NEWER_THAN_2_4,
+    IS_FIRST_AIRFLOW_VERSION,
+)
 
-if airflow.__version__.startswith("1."):
+if IS_FIRST_AIRFLOW_VERSION:
     from airflow.contrib.operators.slack_webhook_operator import SlackWebhookOperator
 else:
     from airflow.providers.slack.operators.slack_webhook import SlackWebhookOperator
@@ -23,8 +26,16 @@ def test_notification_callback_creation():
     assert dag.default_args["on_failure_callback"]
 
 
-@patch("airflow.hooks.base_hook.BaseHook.get_connection")
-@patch("airflow.contrib.operators.slack_webhook_operator.SlackWebhookOperator.__new__")
+@patch(
+    "airflow.hooks.base.BaseHook.get_connection"
+    if IS_AIRFLOW_NEWER_THAN_2_4
+    else "airflow.hooks.base_hook.BaseHook.get_connection"
+)
+@patch(
+    "airflow.contrib.operators.slack_webhook_operator.SlackWebhookOperator.__new__"
+    if IS_FIRST_AIRFLOW_VERSION
+    else "airflow.providers.slack.operators.slack_webhook.SlackWebhookOperator.__new__"
+)
 def test_notification_send(mock_operator_init, mock_get_connection):
     # given
     notifications_config = AirflowDagFactory(