+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+## [0.31.0] - 2023-03-27
+
+### Fixed
+
+-   Use `node_selector` and `container_resources` parameters in `KubernetesPodOperator` if Airflow is 2.3+.
+
 ## [0.30.0] - 2023-02-08
 
 -   Add in_cluster, cluster_context params
@@ -132,7 +138,9 @@ This version brings compatibility with `dbt 1.0`.
 
 -   Initial implementation of `dbt_airflow_manifest_parser` library.
 
-[Unreleased]: https://github.com/getindata/dbt-airflow-factory/compare/0.30.0...HEAD
+[Unreleased]: https://github.com/getindata/dbt-airflow-factory/compare/0.31.0...HEAD
+
+[0.31.0]: https://github.com/getindata/dbt-airflow-factory/compare/0.30.0...0.31.0
 
 [0.30.0]: https://github.com/getindata/dbt-airflow-factory/compare/0.29.0...0.30.0